C#: Move DotNetStub into its own file.
